refactor regression
